Add AbstractN5ImageLoader
